+++ sommor-bundles/src/main/java/com/sommor/bundles/user/model/AuthUser.java
@@ -17,7 +17,7 @@ public class AuthUser {
 
     private String token;
 
-    private Integer expireTime;
+    private Long expireTime;
 
     private Long userId;
